refactor: HeaderDataDispatcher 구현
관심사: 데이터 처리하기
의존: View,Model 의존 하고있음
(이 떄문에 View Model은 아무것도 등록안해도 됨)

역할 데이터를 확인하고 Model에게 템플릿을 받아 View에게
렌더를 탬플릿을 넣어 호출

+++ src/dataDispatcher/HeaderDataDispatcher.js
@@ -1,10 +1,17 @@
 class HedearDataDispatcher {
-  constructor(searchInputToggle, searchMenuToggle) {
+  constructor(
+    searchInputToggle,
+    searchMenuToggle,
+    searchInputView,
+    searchMenuView
+  ) {
     this.historStorage = new Set(
       JSON.parse(localStorage.getItem("localSearchHistory")) // 로컬스토리지 볐다 할수있음
     );
     this.searchInputModel = searchInputToggle;
     this.searchMenuModel = searchMenuToggle;
+    this.searchInputView = searchInputView;
+    this.searchMenuView = searchMenuView;
   }
 
   setData2localStorage(data) {
@@ -16,19 +23,27 @@ class HedearDataDispatcher {
   }
 
   set searchInputData(data) {
-    this.checkdataSize;
-    // 뒤집기
-    // localStorage저장하기
+    if (data.length === 0) {
+      // 나중에 router쪽에서 값이 없으면 차단하는것도 고려
+      return;
+    }
+    const inputData = this.checkdataSize(data);
+    const DOM = this.searchInputModel.getHTML(inputData);
+    this.searchInputView.render(DOM);
   }
 
   get searchInputData() {
-    //
+    const historyData = this.checkHistorySize([...this.historStorage]); // 안해도 될수도 있음
+    const DOM = this.searchInputModel.getHTML(historyData);
+    this.searchInputView.render(DOM);
   }
 
-  set searchMenuData(data) {}
-  get searchMenuData() {
-    this.checkHistorySize;
+  set searchMenuData(data) {
+    const menuDOM = this.searchMenuModel.getHTML(data);
+    this.searchMenuData.render(menuDOM);
   }
+
+  //   get searchMenuData() {}
 
   checkdataSize(...data) {
     const MAX_SIZE = 10;
@@ -44,8 +59,14 @@ class HedearDataDispatcher {
     const historyArr = [...this.historStorage];
     const MAX_SIZE = 10;
 
+    if (historyArr.length === 0) {
+      return ["검색 결과가 없습니다."];
+    }
+
     if (this.historStorage >= MAX_SIZE) {
       this.historStorage.delete(historyArr[0]);
     }
+
+    return [...this.historStorage].reverse();
   }
 }
